+++ AmazonChimeSDKDemo/AmazonChimeSDKDemo/data/JoinMeetingResponse.swift
@@ -11,50 +11,70 @@ import Foundation
 struct CreateMediaPlacementInfo: Codable {
     var audioFallbackUrl: String
     var audioHostUrl: String
-    var turnControlUrl: String
     var signalingUrl: String
+    var turnControlUrl: String
 
     enum CodingKeys: String, CodingKey {
         case audioFallbackUrl = "AudioFallbackUrl"
         case audioHostUrl = "AudioHostUrl"
-        case turnControlUrl = "TurnControlUrl"
         case signalingUrl = "SignalingUrl"
+        case turnControlUrl = "TurnControlUrl"
     }
 }
 
 struct CreateMeetingInfo: Codable {
-    var meetingId: String
+    var externalMeetingId: String
     var mediaPlacement: CreateMediaPlacementInfo
+    var mediaRegion: String
+    var meetingId: String
 
     enum CodingKeys: String, CodingKey {
-        case meetingId = "MeetingId"
+        case externalMeetingId = "ExternalMeetingId"
         case mediaPlacement = "MediaPlacement"
+        case mediaRegion = "MediaRegion"
+        case meetingId = "MeetingId"
     }
 }
 
 struct CreateAttendeeInfo: Codable {
     var attendeeId: String
+    var externalUserId: String
     var joinToken: String
 
     enum CodingKeys: String, CodingKey {
         case attendeeId = "AttendeeId"
+        case externalUserId = "ExternalUserId"
         case joinToken = "JoinToken"
     }
 }
 
-struct CreateJoinInfo: Codable {
-    var title: String
+struct CreateMeeting: Codable {
     var meeting: CreateMeetingInfo
+
+    enum CodingKeys: String, CodingKey {
+        case meeting = "Meeting"
+    }
+}
+
+struct CreateAttendee: Codable {
     var attendee: CreateAttendeeInfo
 
     enum CodingKeys: String, CodingKey {
-        case title = "Title"
+        case attendee = "Attendee"
+    }
+}
+
+struct CreateJoinInfo: Codable {
+    var meeting: CreateMeeting
+    var attendee: CreateAttendee
+
+    enum CodingKeys: String, CodingKey {
         case meeting = "Meeting"
         case attendee = "Attendee"
     }
 }
 
-struct MeetingResponse: Codable {
+struct JoinMeetingResponse: Codable {
     var joinInfo: CreateJoinInfo
 
     enum CodingKeys: String, CodingKey {