+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## [Unreleased]
+### Added
+- **Breaking** Added addtional fields for `CreateAttendeeResponse` and `CreateMeetingResponse`
+- Added `Versioning` class and `sdkVersion` API for retrieving current version of SDK
+
+### Changed
+- Updated demo app to work with updated [amazon-chime-sdk-js serverless demo](https://github.com/aws/amazon-chime-sdk-js/tree/master/demos/serverless). Note that you need to redeploy the serverless demo to work with the updated demo app
 
 ### Changed
 - Changed the default observer caller implementation. Now all the observer callbacks are called on main thread. Make sure to dispatch long-running tasks to another thread to avoid blocking the main thread.